chore(org): reorganize common tests
